feat: Update AppRouter, UsersBlogs, About, Navbar, and BlogDetails components
This commit updates the AppRouter, UsersBlogs, About, Navbar, and BlogDetails components.

In the AppRouter component, the code has been cleaned up by removing unnecessary line breaks.

In the UsersBlogs component, commented out code has been removed and the styling of the Box component has been adjusted.

In the About component, the Box component has been given a maxWidth of 700 and the Typography component has been updated with a white color and a fontSize of 25.

In the Navbar component, the img tag has been modified to remove the marginRight property.

In the BlogDetails component, commented out code has been removed and a new Card component has been added with updated styling and content.

These changes improve the functionality, styling, and user experience of the application.