feat: Update API URL to use production server
This commit updates the API URL in the AuthContext and BlogContext files to use the production server URL instead of the local development server URL. The new URL is 'http://mammedovblog.pythonanywhere.com/'. This change ensures that the application is making requests to the correct server in the production environment.